+++ db/migrate/20141001162851_add_remove_after_to_users.rb
@@ -1,0 +1,5 @@
+class AddRemoveAfterToUsers < ActiveRecord::Migration
+  def change
+  	add_column :users, :remove_after, :datetime
+  end
+end